+++ src/V1/VersionUpgrade/Bradfab/FileUpgrader.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neighborhoods\Buphalo\V1\VersionUpgrade\Beta\V1;
+namespace Neighborhoods\Buphalo\V1\VersionUpgrade\Bradfab;
 
 use LogicException;
 use Neighborhoods\Buphalo\V1\Logger\AwareTrait;
